move type definitions to method implementation files
* check if KeyDirectory is empty before overwriting it in the NewProvider method
