--- relayer/chains/substrate/keystore/types.go
@@ -1,46 +1,0 @@
-package keystore
-
-import (
-	"github.com/99designs/keyring"
-	"github.com/ComposableFi/go-substrate-rpc-client/v4/signature"
-)
-
-type keystore struct {
-	db keyring.Keyring
-}
-
-// Keyring exposes operations over a backend supported by github.com/99designs/keyring.
-type Keyring interface {
-	// List all keys.
-	List() ([]Info, error)
-
-	// Key returns keys by uid respectively.
-	Key(uid string) (Info, error)
-
-	// Delete and DeleteByAddress remove keys from the keyring.
-	Delete(uid string) error
-
-	// NewAccount converts a mnemonic to a private key and BIP-39 HD Path and persists it.
-	// It fails if there is an existing key Info with the same address.
-	NewAccount(name, mnemonic string, network uint8) (Info, error)
-}
-
-// Info is the publicly exposed information about a keypair
-type Info interface {
-	// GetName returns the name of the key
-	GetName() string
-	// GetAddress returns the address as a string
-	GetAddress() string
-	// GetPublicKey returns the publickey as bytes
-	GetPublicKey() []byte
-	// GetKeyringPair returns the keyring pair
-	GetKeyringPair() signature.KeyringPair
-}
-
-// localInfo is the public information about a locally stored key
-// Note: Algo must be last field in struct for backwards amino compatibility
-type localInfo struct {
-	// KeyPair signature.KeyringPair `json:"key_pair"`
-	Name    string                `json:"name"`
-	Keypair signature.KeyringPair `json:"keypair"`
-}
